+++ benches/dynamic_dispatch_benchmark.rs
@@ -11,7 +11,7 @@ const STEP_SIZE: usize = 128usize;
 #[test]
 fn compare_eq() {
     const STEPS: usize = 48000 / STEP_SIZE;
-    let mut output_arr = [[0.0f64; STEP_SIZE]; STEPS];
+    let mut output_arr = [[0.0f32; STEP_SIZE]; STEPS];
 
     let (mut nodes, mut connections) = build_graph();
     for i in 0..STEPS {
@@ -21,12 +21,12 @@ fn compare_eq() {
         .iter()
         .flatten()
         .map(|x| *x)
-        .collect::<Vec<f64>>();
+        .collect::<Vec<f32>>();
     write_wav(&flat, "ng.wav");
-    dbg!(flat.iter().sum::<f64>());
+    dbg!(flat.iter().sum::<f32>());
 }
 
-fn write_wav(samples: &[f64], path: &str) {
+fn write_wav(samples: &[f32], path: &str) {
     let spec = hound::WavSpec {
         channels: 1,
         sample_rate: 48000,
@@ -43,7 +43,7 @@ fn write_wav(samples: &[f64], path: &str) {
 #[bench]
 fn benchmark(b: &mut Bencher) {
     const STEPS: usize = 48000 / STEP_SIZE;
-    let mut output_arr = [[0.0f64; STEP_SIZE]; STEPS];
+    let mut output_arr = [[0.0f32; STEP_SIZE]; STEPS];
 
     //test::black_box({
     //move this into iter for exactly correct output_arr result
@@ -58,16 +58,16 @@ fn benchmark(b: &mut Bencher) {
         .iter()
         .flatten()
         .map(|x| *x)
-        .collect::<Vec<f64>>();
-    dbg!(flat.iter().sum::<f64>());
+        .collect::<Vec<f32>>();
+    dbg!(flat.iter().sum::<f32>());
 }
 
 ///////////////////////////////
 ///////////////////////////////
 ///////////////////////////////
 pub trait Node {
-    fn get_output(&self, name: &str) -> [f64; STEP_SIZE];
-    fn set_input(&mut self, name: &str, val: [f64; STEP_SIZE]);
+    fn get_output(&self, name: &str) -> [f32; STEP_SIZE];
+    fn set_input(&mut self, name: &str, val: [f32; STEP_SIZE]);
     fn process(&mut self);
     fn mark_process(&mut self);
 }
@@ -84,23 +84,23 @@ fn build_graph() -> (HashMap<String, Box<dyn Node>>, Vec<Connection>) {
     nodes.insert(
         "step".to_string(),
         Box::new(Step {
-            step: [-1.0f64; STEP_SIZE], //so we start on 0
+            step: [-1.0f32; STEP_SIZE], //so we start on 0
             need_processes: true,
         }),
     );
     nodes.insert(
         "lfo".to_string(),
         Box::new(Sin {
-            in_val: [0.0f64; STEP_SIZE],
-            out_val: [0.0f64; STEP_SIZE],
+            in_val: [0.0f32; STEP_SIZE],
+            out_val: [0.0f32; STEP_SIZE],
             need_processes: true,
         }),
     );
     nodes.insert(
         "noise".to_string(),
         Box::new(Noise {
-            in_val: [0.0f64; STEP_SIZE],
-            out_val: [0.0f64; STEP_SIZE],
+            in_val: [0.0f32; STEP_SIZE],
+            out_val: [0.0f32; STEP_SIZE],
             need_processes: true,
         }),
     );
@@ -108,9 +108,9 @@ fn build_graph() -> (HashMap<String, Box<dyn Node>>, Vec<Connection>) {
         "highpass".to_string(),
         Box::new(Highpass {
             filter: IIR2::from(IIR2Coefficients::highpass(1000.0, 0.0, 1.0, 48000.0)),
-            in_val: [0.0f64; STEP_SIZE],
-            out_val: [0.0f64; STEP_SIZE],
-            cutoff_val: [0.0f64; STEP_SIZE],
+            in_val: [0.0f32; STEP_SIZE],
+            out_val: [0.0f32; STEP_SIZE],
+            cutoff_val: [0.0f32; STEP_SIZE],
             need_processes: true,
         }),
     );
@@ -118,9 +118,9 @@ fn build_graph() -> (HashMap<String, Box<dyn Node>>, Vec<Connection>) {
         "lowpass".to_string(),
         Box::new(Lowpass {
             filter: IIR2::from(IIR2Coefficients::lowpass(5000.0, 0.0, 1.0, 48000.0)),
-            in_val: [0.0f64; STEP_SIZE],
-            out_val: [0.0f64; STEP_SIZE],
-            cutoff_val: [0.0f64; STEP_SIZE],
+            in_val: [0.0f32; STEP_SIZE],
+            out_val: [0.0f32; STEP_SIZE],
+            cutoff_val: [0.0f32; STEP_SIZE],
             need_processes: true,
         }),
     );
@@ -128,16 +128,16 @@ fn build_graph() -> (HashMap<String, Box<dyn Node>>, Vec<Connection>) {
         "highshelf".to_string(),
         Box::new(Highshelf {
             filter: IIR2::from(IIR2Coefficients::highshelf(2000.0, 6.0, 1.0, 48000.0)),
-            in_val: [0.0f64; STEP_SIZE],
-            out_val: [0.0f64; STEP_SIZE],
-            cutoff_val: [0.0f64; STEP_SIZE],
+            in_val: [0.0f32; STEP_SIZE],
+            out_val: [0.0f32; STEP_SIZE],
+            cutoff_val: [0.0f32; STEP_SIZE],
             need_processes: true,
         }),
     );
     nodes.insert(
         "output".to_string(),
         Box::new(Output {
-            out_val: [0.0f64; STEP_SIZE],
+            out_val: [0.0f32; STEP_SIZE],
         }),
     );
     let connections = vec![
@@ -208,7 +208,7 @@ fn build_graph() -> (HashMap<String, Box<dyn Node>>, Vec<Connection>) {
 fn process_graph_step(
     nodes: &mut HashMap<String, Box<dyn Node>>,
     connections: &mut Vec<Connection>,
-) -> [f64; STEP_SIZE] {
+) -> [f32; STEP_SIZE] {
     for conn in connections {
         let output = {
             if &conn.src_node != "STEPEVENT" {
@@ -217,7 +217,7 @@ fn process_graph_step(
                 src_node.get_output(&conn.src_port)
             } else {
                 //nodes.get_mut(&conn.dst_node).unwrap().process();
-                [0.0f64; STEP_SIZE]
+                [0.0f32; STEP_SIZE]
             }
         };
         nodes
@@ -231,14 +231,14 @@ fn process_graph_step(
     nodes["output"].get_output("audio")
 }
 
-fn rand64(x: f64) -> f64 {
+fn rand64(x: f32) -> f32 {
     // Crappy noise
     ((x * 12000000.9898).sin() * 43758.5453).fract()
 }
 
 pub struct Noise {
-    in_val: [f64; STEP_SIZE],
-    out_val: [f64; STEP_SIZE],
+    in_val: [f32; STEP_SIZE],
+    out_val: [f32; STEP_SIZE],
     need_processes: bool,
 }
 
@@ -255,17 +255,17 @@ impl Node for Noise {
             //dbg!("PROCESS Noise", self.out_val);
         }
     }
-    fn get_output(&self, _name: &str) -> [f64; STEP_SIZE] {
+    fn get_output(&self, _name: &str) -> [f32; STEP_SIZE] {
         self.out_val
     }
-    fn set_input(&mut self, _name: &str, val: [f64; STEP_SIZE]) {
+    fn set_input(&mut self, _name: &str, val: [f32; STEP_SIZE]) {
         self.in_val = val;
     }
 }
 
 pub struct Sin {
-    in_val: [f64; STEP_SIZE],
-    out_val: [f64; STEP_SIZE],
+    in_val: [f32; STEP_SIZE],
+    out_val: [f32; STEP_SIZE],
     need_processes: bool,
 }
 
@@ -282,34 +282,34 @@ impl Node for Sin {
             //dbg!("PROCESS Sin", self.out_val);
         }
     }
-    fn get_output(&self, _name: &str) -> [f64; STEP_SIZE] {
+    fn get_output(&self, _name: &str) -> [f32; STEP_SIZE] {
         self.out_val
     }
-    fn set_input(&mut self, _name: &str, val: [f64; STEP_SIZE]) {
+    fn set_input(&mut self, _name: &str, val: [f32; STEP_SIZE]) {
         self.in_val = val;
     }
 }
 
 pub struct Output {
-    out_val: [f64; STEP_SIZE],
+    out_val: [f32; STEP_SIZE],
 }
 
 impl Node for Output {
     fn mark_process(&mut self) {}
     fn process(&mut self) {}
-    fn get_output(&self, _name: &str) -> [f64; STEP_SIZE] {
+    fn get_output(&self, _name: &str) -> [f32; STEP_SIZE] {
         self.out_val
     }
-    fn set_input(&mut self, _name: &str, val: [f64; STEP_SIZE]) {
+    fn set_input(&mut self, _name: &str, val: [f32; STEP_SIZE]) {
         self.out_val = val;
     }
 }
 
 pub struct Highpass {
     filter: IIR2,
-    out_val: [f64; STEP_SIZE],
-    in_val: [f64; STEP_SIZE],
-    cutoff_val: [f64; STEP_SIZE],
+    out_val: [f32; STEP_SIZE],
+    in_val: [f32; STEP_SIZE],
+    cutoff_val: [f32; STEP_SIZE],
     need_processes: bool,
 }
 
@@ -337,10 +337,10 @@ impl Node for Highpass {
             //dbg!("PROCESS Lowpass", self.out_val);
         }
     }
-    fn get_output(&self, _name: &str) -> [f64; STEP_SIZE] {
+    fn get_output(&self, _name: &str) -> [f32; STEP_SIZE] {
         self.out_val
     }
-    fn set_input(&mut self, name: &str, val: [f64; STEP_SIZE]) {
+    fn set_input(&mut self, name: &str, val: [f32; STEP_SIZE]) {
         if name == "audio" {
             self.in_val = val;
         } else if name == "cutoff" {
@@ -351,9 +351,9 @@ impl Node for Highpass {
 
 pub struct Lowpass {
     filter: IIR2,
-    out_val: [f64; STEP_SIZE],
-    in_val: [f64; STEP_SIZE],
-    cutoff_val: [f64; STEP_SIZE],
+    out_val: [f32; STEP_SIZE],
+    in_val: [f32; STEP_SIZE],
+    cutoff_val: [f32; STEP_SIZE],
     need_processes: bool,
 }
 
@@ -381,10 +381,10 @@ impl Node for Lowpass {
             //dbg!("PROCESS Lowpass", self.out_val);
         }
     }
-    fn get_output(&self, _name: &str) -> [f64; STEP_SIZE] {
+    fn get_output(&self, _name: &str) -> [f32; STEP_SIZE] {
         self.out_val
     }
-    fn set_input(&mut self, name: &str, val: [f64; STEP_SIZE]) {
+    fn set_input(&mut self, name: &str, val: [f32; STEP_SIZE]) {
         if name == "audio" {
             self.in_val = val;
         } else if name == "cutoff" {
@@ -395,9 +395,9 @@ impl Node for Lowpass {
 
 pub struct Highshelf {
     filter: IIR2,
-    out_val: [f64; STEP_SIZE],
-    in_val: [f64; STEP_SIZE],
-    cutoff_val: [f64; STEP_SIZE],
+    out_val: [f32; STEP_SIZE],
+    in_val: [f32; STEP_SIZE],
+    cutoff_val: [f32; STEP_SIZE],
     need_processes: bool,
 }
 
@@ -425,10 +425,10 @@ impl Node for Highshelf {
             //dbg!("PROCESS Lowpass", self.out_val);
         }
     }
-    fn get_output(&self, _name: &str) -> [f64; STEP_SIZE] {
+    fn get_output(&self, _name: &str) -> [f32; STEP_SIZE] {
         self.out_val
     }
-    fn set_input(&mut self, name: &str, val: [f64; STEP_SIZE]) {
+    fn set_input(&mut self, name: &str, val: [f32; STEP_SIZE]) {
         if name == "audio" {
             self.in_val = val;
         } else if name == "cutoff" {
@@ -438,7 +438,7 @@ impl Node for Highshelf {
 }
 
 pub struct Step {
-    step: [f64; STEP_SIZE],
+    step: [f32; STEP_SIZE],
     need_processes: bool,
 }
 
@@ -457,8 +457,8 @@ impl Node for Step {
             //dbg!("PROCESS Step", self.step);
         }
     }
-    fn get_output(&self, _name: &str) -> [f64; STEP_SIZE] {
+    fn get_output(&self, _name: &str) -> [f32; STEP_SIZE] {
         self.step
     }
-    fn set_input(&mut self, _name: &str, _val: [f64; STEP_SIZE]) {}
+    fn set_input(&mut self, _name: &str, _val: [f32; STEP_SIZE]) {}
 }
